Added params check functions to utils
